fix for float to int conversion in fractals

+++ gprMax/fractals_generate_ext.pyx
@@ -21,7 +21,7 @@ cimport numpy as np
 from cython.parallel import prange
 
 
-cpdef void generate_fractal2D(int nx, int ny, int nthreads, int b, np.float64_t[:] weighting, np.float64_t[:] v1, np.complex128_t[:, ::1] A, np.complex128_t[:, ::1] fractalsurface):
+cpdef void generate_fractal2D(int nx, int ny, int nthreads, float b, np.float64_t[:] weighting, np.float64_t[:] v1, np.complex128_t[:, ::1] A, np.complex128_t[:, ::1] fractalsurface):
     """This function generates a fractal surface for a 2D array.
         
     Args:
@@ -53,7 +53,7 @@ cpdef void generate_fractal2D(int nx, int ny, int nthreads, int b, np.float64_t[
                 fractalsurface[i, j] = A[i, j] / B
 
 
-cpdef void generate_fractal3D(int nx, int ny, int nz, int nthreads, int b, np.float64_t[:] weighting, np.float64_t[:] v1, np.complex128_t[:, :, ::1] A, np.complex128_t[:, :, ::1] fractalvolume):
+cpdef void generate_fractal3D(int nx, int ny, int nz, int nthreads, float b, np.float64_t[:] weighting, np.float64_t[:] v1, np.complex128_t[:, :, ::1] A, np.complex128_t[:, :, ::1] fractalvolume):
     """This function generates a fractal volume for a 3D array.
 
     Args:
